refactor #73: move directory
