+++ API-Server/src/main/java/swm/hkcc/LGTM/app/modules/registration/domain/additionalInfoProvider/AdditionalInfoProvider.java
@@ -1,4 +1,4 @@
-package swm.hkcc.LGTM.app.modules.registration.service.additionalInfoProvider;
+package swm.hkcc.LGTM.app.modules.registration.domain.additionalInfoProvider;
 
 import swm.hkcc.LGTM.app.modules.member.domain.Member;
 import swm.hkcc.LGTM.app.modules.registration.dto.registrationSeniorDetailResponse.AdditionalInfo;
