+++ API-Server/src/test/java/swm/hkcc/LGTM/app/modules/registration/service/SeniorEnrollDetailServiceTest.java
@@ -25,7 +25,7 @@ import swm.hkcc.LGTM.app.modules.registration.dto.registrationSeniorDetailRespon
 import swm.hkcc.LGTM.app.modules.registration.dto.registrationSeniorDetailResponse.RegistrationSeniorDetailResponse;
 import swm.hkcc.LGTM.app.modules.registration.repository.MissionHistoryRepository;
 import swm.hkcc.LGTM.app.modules.registration.repository.MissionRegistrationRepository;
-import swm.hkcc.LGTM.app.modules.registration.service.additionalInfoProvider.AdditionalInfoProviderFactory;
+import swm.hkcc.LGTM.app.modules.registration.domain.additionalInfoProvider.AdditionalInfoProviderFactory;
 import swm.hkcc.LGTM.app.modules.review.domain.Review;
 import swm.hkcc.LGTM.app.modules.review.repository.ReviewRepository;
